+++ src/pyshotter/__init__.py
@@ -28,7 +28,7 @@ from .features import (
     ScreenshotHistory,
 )
 
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 __author__ = "Abdoullah Ndao"
 __email__ = "abdoullahaljersi@gmail.com"
 __date__ = "2024"